Move OID::Spatial to separate file: oid/spatial.rb
